chore: split user worker tests to multiple files
