+++ crates/base/tests/oak_user_worker_tests.rs
@@ -1,11 +1,11 @@
-mod null_body_status_null_body_tests;
-mod tls_invalid_data_tests;
-
 use base::worker_ctx::{create_worker, WorkerRequestMsg};
 use hyper::{Body, Request, Response};
 use sb_worker_context::essentials::{EdgeContextInitOpts, EdgeContextOpts, EdgeUserRuntimeOpts};
 use std::collections::HashMap;
 use tokio::sync::oneshot;
+
+// NOTE: Only add user worker tests that's using oak server here.
+// Any other user worker tests, add to `user_worker_tests.rs`.
 
 #[tokio::test]
 async fn test_oak_server() {
@@ -77,34 +77,4 @@ async fn test_file_upload() {
     let body_bytes = hyper::body::to_bytes(res.into_body()).await.unwrap();
 
     assert_eq!(body_bytes, "file-type: text/plain");
-}
-
-#[tokio::test]
-async fn test_json_imports() {
-    let user_rt_opts = EdgeUserRuntimeOpts::default();
-    let opts = EdgeContextInitOpts {
-        service_path: "./test_cases/json_import".into(),
-        no_module_cache: false,
-        import_map_path: None,
-        env_vars: HashMap::new(),
-        conf: EdgeContextOpts::UserWorker(user_rt_opts),
-    };
-    let worker_req_tx = create_worker(opts).await.unwrap();
-    let (res_tx, res_rx) = oneshot::channel::<Result<Response<Body>, hyper::Error>>();
-
-    let req = Request::builder()
-        .uri("/")
-        .method("GET")
-        .body(Body::empty())
-        .unwrap();
-
-    let msg = WorkerRequestMsg { req, res_tx };
-    let _ = worker_req_tx.send(msg);
-
-    let res = res_rx.await.unwrap().unwrap();
-    assert!(res.status().as_u16() == 200);
-
-    let body_bytes = hyper::body::to_bytes(res.into_body()).await.unwrap();
-
-    assert_eq!(body_bytes, r#"{"version":"1.0.0"}"#);
 }
